# rafthttp: introduce msgappv2 stream format