msgappv2 stream is used to send all MsgApp, and replaces the
functionality of msgapp stream. Compared to v1, it has several
advantanges:
1. The output message is exactly the same with the input one, which
cannot be done in v1.
2. It uses one connection to stream persistently, which prevents message
reorder and saves the time to request stream.
3. It transmits 10 addiontional bytes in the procedure of committing one
proposal, which is trivia for idle time.
4. It transmits much less bytes when committing mutliple proposals or keep
committing proposals.

Benchmark performance:
before:
BenchmarkSendingMsgApp	 100000	    15938 ns/op	  4.02 MB/s	    848 B/op	     19 allocs/op
after:
BenchmarkSendingMsgApp	  100000	     17834 ns/op	   3.59 MB/s	    1337 B/op	      28 allocs/op

It is caused by code inefficiency, and can be improved to the level that is better than v1. This will be done in the follow-up.

This fixes the problem #2153 that proposal on follower takes 0.1s to finish even on local machine. It improves the latency to commit proposal in low load at *100x* times. The details are listed in the following comments.

/cc @barakmich