@yichengq 2015-04-03
Based on discussion OOL with xiang, use format "type length data" instead of raftpb.Entry does save 16 bytes for us, but it increase much complicity and becomes hard to explain. Will change to use "length raftpb.Entry".