@xiang90 2015-04-03
@yichengq I do not quite care about MB and allocation in this pull request. We are improving the latency under low load, right? So we need to show that numbers.