@yichengq 2015-04-03
Sure.

The latency under low load before:
```
Response time histogram:
  0.031 [1]	|
  0.038 [0]	|
  0.045 [0]	|
  0.052 [0]	|
  0.060 [0]	|
  0.067 [0]	|
  0.074 [0]	|
  0.082 [0]	|
  0.089 [0]	|
  0.096 [0]	|
  0.103 [99]	|∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎
```

after this PR:
```
Response time histogram:
  0.001 [1]	|
  0.001 [75]	|∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎
  0.001 [17]	|∎∎∎∎∎∎∎∎∎
  0.001 [2]	|∎
  0.001 [3]	|∎
  0.002 [1]	|
  0.002 [0]	|
  0.002 [0]	|
  0.002 [0]	|
  0.002 [0]	|
  0.002 [1]	|
```
100x improvements :wink: